requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Werkzeug==2.3.7
requests
google-auth
google-cloud-storage